Adds quota details in RateLimitException
