+++ tests/Auth0.Core.UnitTests/RateLimitDeserializationTests.cs
@@ -23,7 +23,8 @@ namespace Auth0.Core.UnitTests
 
     public class RateLimitDeserializationData : IEnumerable<object[]>
     {
-        private static HttpHeaders CreateHeaders(int? limit, int? remaining, long? reset)
+        private static HttpHeaders CreateHeaders(
+            int? limit, int? remaining, long? reset, string clientQuota = null, string orgQuota = null)
         {
             var client = new HttpRequestMessage(HttpMethod.Get, "https://fake");
             if (limit != null)
@@ -32,34 +33,99 @@ namespace Auth0.Core.UnitTests
                 client.Headers.Add("x-ratelimit-remaining", remaining.ToString());
             if (reset != null)
                 client.Headers.Add("x-ratelimit-reset", reset.ToString());
+            if (clientQuota != null)
+                client.Headers.Add("X-Quota-Client-Limit", clientQuota);
+            if (orgQuota != null)
+                client.Headers.Add("X-Quota-Organization-Limit", orgQuota);
             return client.Headers;
         }
 
         public IEnumerator<object[]> GetEnumerator()
         {
-            yield return new object[] { CreateHeaders(100, 10, 1585694338),
+            yield return new object[]
+            {
+                CreateHeaders(100, 10, 1585694338),
                 new RateLimit
                 {
                     Limit = 100,
                     Remaining = 10,
                     Reset = new DateTimeOffset(2020, 3, 31, 22, 38, 58, TimeSpan.Zero)
-                } };
+                }
+            };
 
-            yield return new object[] { CreateHeaders(5, 100, 1585694338),
+            yield return new object[]
+            {
+                CreateHeaders(5, 100, 1585694338),
                 new RateLimit
                 {
                     Limit = 5,
                     Remaining = 100,
                     Reset = new DateTimeOffset(2020, 3, 31, 22, 38, 58, TimeSpan.Zero)
-                } };
+                }
+            };
 
-            yield return new object[] { CreateHeaders(null, 10, null),
+            yield return new object[]
+            {
+                CreateHeaders(null, 10, null),
                 new RateLimit
                 {
                     Limit = 0,
                     Remaining = 10,
                     Reset = null
-                } };
+                }
+            };
+            yield return new object[]
+            {
+                CreateHeaders(null, 10, null,
+                    "b=per_hour;q=10;r=9;t=774,b=per_day;q=100;r=99;t=22374", null),
+                new RateLimit
+                {
+                    Limit = 0,
+                    Remaining = 10,
+                    Reset = null,
+                    ClientQuotaLimit = new ClientQuotaLimit()
+                    {
+                        PerHour = new QuotaLimit
+                        {
+                            Quota = 10,
+                            Remaining = 9,
+                            Time = 774
+                        },
+                        PerDay = new QuotaLimit
+                        {
+                            Quota = 100,
+                            Remaining = 99,
+                            Time = 22374
+                        }
+                    }
+                }
+            };
+            yield return new object[]
+            {
+                CreateHeaders(null, 10, null,
+                    null, "b=per_hour;q=10;r=9;t=774,b=per_day;q=100;r=99;t=22374"),
+                new RateLimit
+                {
+                    Limit = 0,
+                    Remaining = 10,
+                    Reset = null,
+                    OrganizationQuotaLimit = new OrganizationQuotaLimit()
+                    {
+                        PerHour = new QuotaLimit
+                        {
+                            Quota = 10,
+                            Remaining = 9,
+                            Time = 774
+                        },
+                        PerDay = new QuotaLimit
+                        {
+                            Quota = 100,
+                            Remaining = 99,
+                            Time = 22374
+                        }
+                    }
+                }
+            };
         }
 
         IEnumerator IEnumerable.GetEnumerator()
